Merge branch 'sctp-remove-typedefs-from-structures-part-3'
Xin Long says:

====================
sctp: remove typedefs from structures part 3

As we know, typedef is suggested not to use in kernel, even checkpatch.pl
also gives warnings about it. Now sctp is using it for many structures.

All this kind of typedef's using should be removed. This patchset is the
part 3 to remove it for another 11 basic structures from linux/sctp.h.

Just as the part 1 and 2, No any code's logic would be changed in these
patches, only cleaning up.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 6a9bef4eee6b3a2edbe43483695d1936b7acd3dd ]