Merge branch 'netvsc-minor-fixes-and-optimization'
Stephen Hemminger says:

====================
netvsc: minor fixes and optimization

This is a subset of earlier submission with a few more fixes
found during testing. The are two small optimizations, one is to
better manage the receive completion ring, and the other is removing
one unneeded level of indirection.

Will submit the improved VF support and buffer sizing in a later
patch so they get more review.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: fe21b269597d9d5bca4e2c0cebe2bdf658f6289e ]