ARM: dts: keystone-k2g-evm: Add and enable DSP CMA memory pool
A CMA memory pool reserved memory node is added, and is attached
to the DSP node through the 'memory-region' property on the K2G
EVM board. This area will be used for allocating virtio rings and
buffers. This node allows the DSP Memory Protection and Address
Extension (MPAX) module to be configured properly for the DSP
processor, and matches the values used on the other Keystone 2
boards for software compatibility.

The reserved memory node and the user DSP node are also marked
okay to enable the DSP on the 66AK2G EVM board.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Santosh Shilimkar <ssantosh@kernel.org>

[ upstream commit: 786d7114d25ea990792bc4cb9a0e725a7ac78f81 ]