Merge branch 'nfp-process-MTU-updates-from-firmware-flower-app'
Simon Horman says:

====================
nfp: process MTU updates from firmware flower app

The first patch of this series moves processing of control messages from a
BH handler to a workqueue. That change makes it safe to process MTU
updates from the firmware which is added by the second patch of this
series.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 7440cccee02a72a2e16f26c7cff674c15f5cd85b ]