Merge branch 'phylib-xcvr-type'
Florian Fainelli says:

====================
net: Bring back transceiver type for PHYLIB

With the introduction of the xLINKSETTINGS ethtool APIs, the transceiver type
was deprecated, but in that process we lost some useful information that PHYLIB
was consistently reporting about internal vs. external PHYs.

This brings back transceiver as a read-only field that is only consumed in the
legacy path where ETHTOOL_GET is called but the underlying drivers implement the
new style klink_settings API.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 59ec693a45b9bc0a3977552bc3421202fd7fb2f1 ]