Merge branch 'nfp-fix-ethtool-stats-and-page-allocation'
Jakub Kicinski says:

====================
nfp: fix ethtool stats and page allocation

Two fixes for net.  First one makes sure we handle gather of stats on
32bit machines correctly (ouch).  The second fix solves a potential
NULL-deref if we fail to allocate a page with XDP running.

I used Fixes: tags pointing to where the bug was introduced, but for
patch 1 it has been in the driver "for ever" and fix won't backport
cleanly beyond commit 325945ede6d4 ("nfp: split software and hardware
vNIC statistics") which is in net.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: befc7a3324b45b7478b29861f2a77f8b2bc17922 ]